+++ src/langcheck/metrics/__init__.py
@@ -4,6 +4,10 @@ from langcheck.metrics.custom_text_quality import (
     custom_pairwise_evaluator,
 )
 from langcheck.metrics.en.pairwise_text_quality import pairwise_comparison
+from langcheck.metrics.en.query_based_text_quality import (
+    answer_relevance,
+    answer_safety,
+)
 from langcheck.metrics.en.reference_based_text_quality import (
     answer_correctness,
     rouge1,
@@ -13,7 +17,6 @@ from langcheck.metrics.en.reference_based_text_quality import (
 )
 from langcheck.metrics.en.reference_free_text_quality import (
     ai_disclaimer_similarity,
-    answer_relevance,
     flesch_kincaid_grade,
     flesch_reading_ease,
     fluency,
@@ -42,6 +45,7 @@ __all__ = [
     "ai_disclaimer_similarity",
     "answer_correctness",
     "answer_relevance",
+    "answer_safety",
     "contains_all_strings",
     "contains_any_strings",
     "contains_regex",
