+++ src/langcheck/metrics/en/__init__.py
@@ -1,4 +1,8 @@
 from langcheck.metrics.en.pairwise_text_quality import pairwise_comparison
+from langcheck.metrics.en.query_based_text_quality import (
+    answer_relevance,
+    answer_safety,
+)
 from langcheck.metrics.en.reference_based_text_quality import (
     answer_correctness,
     rouge1,
@@ -8,7 +12,6 @@ from langcheck.metrics.en.reference_based_text_quality import (
 )
 from langcheck.metrics.en.reference_free_text_quality import (
     ai_disclaimer_similarity,
-    answer_relevance,
     flesch_kincaid_grade,
     flesch_reading_ease,
     fluency,
@@ -24,6 +27,7 @@ __all__ = [
     "ai_disclaimer_similarity",
     "answer_correctness",
     "answer_relevance",
+    "answer_safety",
     "context_relevance",
     "factual_consistency",
     "flesch_kincaid_grade",
