+++ src/langcheck/metrics/ja/__init__.py
@@ -1,5 +1,9 @@
 from langcheck.metrics.ja._tokenizers import JanomeTokenizer, MeCabTokenizer
 from langcheck.metrics.ja.pairwise_text_quality import pairwise_comparison
+from langcheck.metrics.ja.query_based_text_quality import (
+    answer_relevance,
+    answer_safety,
+)
 from langcheck.metrics.ja.reference_based_text_quality import (
     answer_correctness,
     rouge1,
@@ -8,7 +12,6 @@ from langcheck.metrics.ja.reference_based_text_quality import (
     semantic_similarity,
 )
 from langcheck.metrics.ja.reference_free_text_quality import (
-    answer_relevance,
     fluency,
     sentiment,
     tateishi_ono_yamada_reading_ease,
@@ -22,6 +25,7 @@ from langcheck.metrics.ja.source_based_text_quality import (
 __all__ = [
     "answer_correctness",
     "answer_relevance",
+    "answer_safety",
     "context_relevance",
     "factual_consistency",
     "JanomeTokenizer",
