+++ src/langcheck/metrics/ja/reference_free_text_quality.py
@@ -5,7 +5,6 @@ from typing import List, Optional, Tuple
 import regex as re
 
 from langcheck.metrics._validation import (
-    validate_parameters_answer_relevance,
     validate_parameters_reference_free,
 )
 from langcheck.metrics.eval_clients import EvalClient
@@ -528,56 +527,6 @@ def tateishi_ono_yamada_reading_ease(
         reference_outputs=None,
         sources=None,
         explanations=None,
-        metric_values=scores,
-        language="ja",
-    )
-
-
-def answer_relevance(
-    generated_outputs: List[str] | str,
-    prompts: List[str] | str,
-    eval_model: EvalClient,
-) -> MetricValue[Optional[float]]:
-    """Calculates the relevance of generated outputs to the prompt. This metric
-    takes on float values of either 0.0 (Not Relevant), 0.5 (Partially
-    Relevant), or 1.0 (Fully Relevant). The score may also be `None` if it could
-    not be computed.
-
-    We currently only support the evaluation based on an EvalClient.
-    """
-    generated_outputs, prompts = validate_parameters_answer_relevance(
-        generated_outputs, prompts
-    )
-
-    answer_relevance_template = eval_model.load_prompt_template(
-        language="ja", metric_name="answer_relevance"
-    )
-
-    populated_prompts = [
-        answer_relevance_template.render(
-            {"gen_output": gen_output, "user_query": prompt}
-        )
-        for gen_output, prompt in zip(generated_outputs, prompts)
-    ]
-
-    scores, explanations = eval_model.get_score(
-        metric_name="answer relevance",
-        language="ja",
-        prompts=populated_prompts,
-        score_map={
-            "Not Relevant": 0.0,
-            "Partially Relevant": 0.5,
-            "Fully Relevant": 1.0,
-        },
-    )
-
-    return MetricValue(
-        metric_name="answer_relevance",
-        prompts=prompts,
-        generated_outputs=generated_outputs,
-        reference_outputs=None,
-        sources=None,
-        explanations=explanations,
         metric_values=scores,
         language="ja",
     )
