+++ tests/metrics/en/test_reference_free_text_quality.py
@@ -4,7 +4,6 @@ from unittest.mock import Mock, patch
 import pytest
 from langcheck.metrics.en import (
     ai_disclaimer_similarity,
-    answer_relevance,
     flesch_kincaid_grade,
     flesch_reading_ease,
     fluency,
@@ -243,32 +242,3 @@ def test_ai_disclaimer_similarity_openai(generated_outputs):
         # the AI disclaimer phrase, the AI disclaimer language similarity should
         # be 1.
         assert 0.99 <= metric_value <= 1
-
-
-@pytest.mark.parametrize(
-    "generated_outputs,prompts",
-    [
-        ("Tokyo is Japan's capital city.", "What is the capital of Japan?"),
-        (["Tokyo is Japan's capital city."], ["What is the capital of Japan?"]),
-    ],
-)
-def test_answer_relevance_eval_client(generated_outputs, prompts):
-    eval_client = MockEvalClient()
-    metric_value = answer_relevance(
-        generated_outputs, prompts, eval_model=eval_client
-    )
-    # MockEvalClient without any argument returns None
-    assert metric_value.metric_values[0] is None
-
-    answer_relevance_assessment_to_score = {
-        "Not Relevant": 0.0,
-        "Partially Relevant": 0.5,
-        "Fully Relevant": 1.0,
-    }
-
-    for option in answer_relevance_assessment_to_score:
-        eval_client = MockEvalClient(option)
-        metric_value = answer_relevance(
-            generated_outputs, prompts, eval_model=eval_client
-        )
-        assert metric_value == answer_relevance_assessment_to_score[option]
